Reorganized HeatMap Functions into HeatMap.kt
